finished all backedn work, fixed router bugs and associations

+++ Router/campusRouter.js
@@ -11,13 +11,22 @@ router.get('/', (req,res) =>{
 })
 
 //GET localhost:3000/api/campuses/1 
-router.get('/:id', (req,res) =>{
-	Campus.findByPk(req.params.id)
-	.then(campus=> res.status(200).send(campus))
-	.catch(err => console.log(err))
+router.get('/:id', async(req,res) =>{
+	await Campus.findByPk(req.params.id)
+	.then(campus =>{
+		if(campus !== null)
+		{
+			res.status(200).send(campus);
+		}
+		else
+		{
+			res.status(404).send();
+		}
+	})
 })
 
-//POST localhost:3000/api/campuses //issues with duplicates
+
+//POST localhost:3000/api/campuses
 router.post('/', async (req,res)=>{
 	let data = req.body;
 	await Campus.create({
@@ -34,31 +43,61 @@ router.post('/', async (req,res)=>{
 
 //PUT localhost:3000/api/campuses/1
 router.put('/:id', async (req,res)=>{
-	let data = req.body;
-	await Campus.update(	//information to update it with
+	await Campus.findByPk(req.params.id)
+	.then(async campus =>{
+		if(campus != null)
 		{
-			name: data.name,
-			bio: data.bio,
-			address: data.address,
-			img: data.img
-		},
-		{where: { id: req.params.id}}		//location in the database to update
-	)
+			let data = req.body;
+			await Campus.update(	//information to update it with
+			{
+				name: data.name,
+				bio: data.bio,
+				address: data.address,
+				img: data.img
+			},
+			{where: { id: req.params.id}}		//location in the database to update
+			)
 		
-	console.log("entry has been updated");
-	await Campus.findAll({ include: [Student] })
-	.then(campuses => res.status(201).json(campuses))
-	.catch(err => console.log(err))
+			console.log("entry has been updated");
+			await Campus.findAll({ include: [Student] })
+			.then(campuses => res.status(201).json(campuses))
+			.catch(err => console.log(err))
+		}
+		else{
+			let data = req.body;
+			await Campus.create({
+				name: data.name,
+				bio: data.bio,
+				address: data.address,
+				img: data.img
+			});
+			console.log('User added');
+
+			Campus.findAll({ include: [Student] }) //Eager Loading
+			.then(campuses => res.status(201).json(campuses))
+		}
+	
+	})
 })
 
 //DELETE localhost:3000/api/campuses/1
 router.delete('/:id', async(req,res)=>{
-	await Campus.destroy({ where: {id : req.params.id}});
-	console.log("entry DESTORYED");
+	await Campus.findByPk(req.params.id)
+	.then(async campus => {
+		if(campus != null)
+		{
+			await Campus.destroy({ where: {id : req.params.id}});
+			console.log("entry DESTORYED");
 
-	await Campus.findAll({ include: [Student] })
-		.then(campuses => res.status(201).json(campuses))
-		.catch(err => console.log(err))
+			await Campus.findAll({ include: [Student] })
+			.then(campuses => res.status(201).json(campuses))
+			.catch(err => console.log(err))
+		}
+		else{
+			res.status(404).send();
+		}
+	})
+	
 })
 
 module.exports = router;
